+++ src/main/java/es/sd/practica1/RepositorioModelos.java
@@ -8,7 +8,7 @@ public interface RepositorioModelos extends JpaRepository<Modelos, Integer> {
 	Modelos findByNombre (String nombre);
 	Modelos findByValorFacial (double valorFacial);
 	Modelos findByUnidad (String unidad);
-	Modelos findByDiametro (int diametro);
+	Modelos findByDiametro (double diametro);
 	Modelos findByPeso (double peso);
 	Modelos findByMetal (String metal);
 	Modelos findByDescripcion (String descripcion);
